Moved UpdateCampaignContact() from ContactExtensions to EditCampaignCommandHandler
