+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Campaigns/EditCampaignCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using AllReady.Areas.Admin.Models;
 using AllReady.Extensions;
@@ -44,7 +46,7 @@ namespace AllReady.Areas.Admin.Features.Campaigns
             campaign.ManagingOrganizationId = message.Campaign.OrganizationId;
             campaign.ImageUrl = message.Campaign.ImageUrl;
 
-            campaign = campaign.UpdateCampaignContact(message.Campaign, _context);
+            CreateUpdateOrDeleteCampaignPrimaryContact(campaign, message.Campaign);
             campaign.CampaignImpact = campaign.CampaignImpact.UpdateModel(message.Campaign.CampaignImpact);
             campaign.Location = campaign.Location.UpdateModel(message.Campaign.Location);
 
@@ -66,6 +68,53 @@ namespace AllReady.Areas.Admin.Features.Campaigns
             await _context.SaveChangesAsync().ConfigureAwait(false);
 
             return campaign.Id;
+        }
+
+        void CreateUpdateOrDeleteCampaignPrimaryContact(Campaign campaign, IPrimaryContactModel contactModel)
+        {
+            var hasPrimaryContact = campaign.CampaignContacts != null &&
+                campaign.CampaignContacts.Any(campaignContact => campaignContact.ContactType == (int)ContactTypes.Primary);
+
+            var addOrUpdatePrimaryContact = !string.IsNullOrWhiteSpace(string.Concat(contactModel.PrimaryContactEmail?.Trim() + contactModel.PrimaryContactFirstName?.Trim(), contactModel.PrimaryContactLastName?.Trim(), contactModel.PrimaryContactPhoneNumber?.Trim()));
+
+            // Update existing Primary Campaign Contact
+            if (hasPrimaryContact && addOrUpdatePrimaryContact)
+            {
+                var contactId = campaign.CampaignContacts.Single(campaignContact => campaignContact.ContactType == (int)ContactTypes.Primary).ContactId;
+                var contact = _context.Contacts.Single(c => c.Id == contactId);
+
+                contact.Email = contactModel.PrimaryContactEmail;
+                contact.FirstName = contactModel.PrimaryContactFirstName;
+                contact.LastName = contactModel.PrimaryContactLastName;
+                contact.PhoneNumber = contactModel.PrimaryContactPhoneNumber;
+            }
+            // Delete existing Primary Campaign Contact
+            else if (hasPrimaryContact && !addOrUpdatePrimaryContact)
+            {
+                var campaignContact = campaign.CampaignContacts.Single(cc => cc.ContactType == (int)ContactTypes.Primary);
+                var contact = _context.Contacts.Single(c => c.Id == campaignContact.ContactId);
+                _context.Remove(contact);
+                _context.Remove(campaignContact);
+            }
+            // Add a Primary Campaign Contact
+            else if (!hasPrimaryContact && addOrUpdatePrimaryContact)
+            {
+                var campaignContact = new CampaignContact()
+                {
+                    ContactType = (int)ContactTypes.Primary,
+                    Contact = new Contact()
+                    {
+                        Email = contactModel.PrimaryContactEmail,
+                        FirstName = contactModel.PrimaryContactFirstName,
+                        LastName = contactModel.PrimaryContactLastName,
+                        PhoneNumber = contactModel.PrimaryContactPhoneNumber
+                    },
+                };
+
+                if (campaign.CampaignContacts == null) campaign.CampaignContacts = new List<CampaignContact>();
+                campaign.CampaignContacts.Add(campaignContact);
+                _context.Add(campaignContact);
+            }
         }
     }
 }
